requirements.txt
celery==5.2.7
Django==4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-celery-beat==2.4.0
django-filter==22.1
drf-spectacular==0.25.1
factory-boy==3.2.1
google-cloud-secret-manager==2.15.0
google-cloud-tasks==2.12.1
gunicorn==20.1.0
pre-commit==2.21.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
whitenoise==6.3.0
